requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.2.0
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
celery==5.4.0
cffi==1.17.1
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
excel==1.0.1
Flask==3.0.3
Flask-Cache==0.13.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Excel==0.0.7
Flask-Ext==0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.1
greenlet==3.1.0
html5lib==1.1
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
kombu==5.4.1
lml==0.1.0
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
passlib==1.7.4
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.1.4
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
six==1.16.0
SQLAlchemy==2.0.34
texttable==1.7.0
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
xlrd==2.0.1
zipp==3.20.1
zopfli==0.2.3
Flask-Login>=0.6.2
